split main controller into multiple
