--- src/main/java/mops/gruppen2/controller/MopsController.java
@@ -1,29 +1,0 @@
-package mops.gruppen2.controller;
-
-import mops.gruppen2.domain.exception.EventException;
-import mops.gruppen2.domain.exception.PageNotFoundException;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-
-@Controller
-public class MopsController {
-
-    @GetMapping("")
-    public String redirect() {
-        return "redirect:/gruppen2/";
-    }
-
-    @GetMapping("/logout")
-    public String logout(HttpServletRequest request) throws ServletException {
-        request.logout();
-        return "redirect:/gruppen2/";
-    }
-
-    @GetMapping("*")
-    public String defaultLink() throws EventException {
-        throw new PageNotFoundException("\uD83D\uDE41");
-    }
-}
